requirements.txt
pydantic_settings==2.11.0
PyYAML==6.0.3
Requests==2.32.5
urllib3==2.5.0
